refactor: OGP用Google Fontsの取得を汎用化しJetBrains Monoを追加
M PLUS 2専用だったフォント取得処理を、family/weightを引数に取る
getGoogleFont として切り出す。getMPlus2Font はその薄いラッパーに
なり、呼び出し側のインターフェースは変わらない。

サブセット指定の文字列は重複排除・ソートしてからURLに載せるよう
にし、同じグリフ集合でキャッシュが効きやすくした。あわせてOGP画
像のコード表示で使うJetBrains Monoのローダーを追加する。
