+++ apps/main/src/shared/og/get-jetbrains-mono-font.ts
@@ -1,0 +1,9 @@
+import { getGoogleFont } from './get-google-font';
+
+export function getJetBrainsMonoFont({
+  text,
+}: {
+  text: string;
+}): Promise<ArrayBuffer> {
+  return getGoogleFont({ family: 'JetBrains Mono', weight: 400, text });
+}
